Hello @lavrton,
Please review my pull request. I've added a "Save File" button with the option to customize the file name. I found that every time we save a file in Polotno Studio, it's automatically named "polotno.json." I've worked on changing this behavior to allow for custom file names. Your feedback on this change is greatly appreciated.

Thank you for your time and assistance!